requirements.txt
# Logs Ingestion Pipeline (Without Spark)
pydantic
python-dotenv
fastembed
qdrant-client
clickhouse-connect
